Move helpers to Service
